+++ plugin/src/main/kotlin/org/jlleitschuh/gradle/ktlint/TaskCreation.kt
@@ -1,6 +1,5 @@
 package org.jlleitschuh.gradle.ktlint
 
-import org.gradle.api.Project
 import org.gradle.api.file.FileTree
 import org.gradle.api.tasks.TaskCollection
 import org.gradle.api.tasks.TaskProvider
@@ -37,16 +36,6 @@ internal fun createFormatTask(
         PatternSet()
     ) {
         mustRunAfter(project.tasks.named(KtLintFormatTask.KOTLIN_SCRIPT_TASK_NAME))
-
-        val rootProjectName = project.rootProject.name
-        var parentProject: Project? = project.parent
-        while (parentProject != null && parentProject.name != rootProjectName) {
-            val parentProjectPath = parentProject.path
-            parentProject.plugins.withId("org.jlleitschuh.gradle.ktlint") {
-                mustRunAfter("$parentProjectPath:${KtLintFormatTask.KOTLIN_SCRIPT_TASK_NAME}")
-            }
-            parentProject = parentProject.parent
-        }
 
         description = KtLintFormatTask.buildDescription(".kt")
         configureBaseCheckTask(pluginHolder) {