+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [Unreleased]
+
+## [14.2.0] - 2026-03-12
+
 - Make plugin compatible with Isolated Projects [#1032](https://github.com/JLLeitschuh/ktlint-gradle/pull/1032)
 
 ## [14.1.0] - 2026-03-05
@@ -794,7 +797,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 - Renamed task with name `ktlint` to `ktlintCheck` (#3)
 - Renamed tasks with names `ktlint[source set name]` to `ktlint[source set name]Check` (#3)
 
-[unreleased]: https://github.com/JLLeitschuh/ktlint-gradle/compare/v14.1.0...HEAD
+[unreleased]: https://github.com/JLLeitschuh/ktlint-gradle/compare/v14.2.0...HEAD
+[14.2.0]: https://github.com/JLLeitschuh/ktlint-gradle/compare/v14.1.0...v14.2.0
 [14.1.0]: https://github.com/JLLeitschuh/ktlint-gradle/compare/v14.0.1...v14.1.0
 [14.0.1]: https://github.com/JLLeitschuh/ktlint-gradle/compare/v14.0.0...v14.0.1
 [14.0.0]: https://github.com/JLLeitschuh/ktlint-gradle/compare/v13.1.0...v14.0.0